refactor: move MetricSampleSizeDisplay into  src/components/features/experiments/
